[Feat] 편지 보따리 달성도 데이터 변경 - 퍼센트 25퍼센트 단위가 아닌 실수형으로 바로바로 반영으로 변경됨.

+++ src/main/java/com/likelion/allForOne/domain/tblGroup/dto/GroupResponseDto.java
@@ -22,7 +22,7 @@ public class GroupResponseDto {
         private boolean ownerYn;                                //유저의 방장 여부
         private String groupName;                               //방(그룹)이름
         private long dayAfterCnt;                               //생성일로부터 며칠
-        private int achievePercent;                             //보따리 달성 퍼센트
+        private double achievePercent;                          //보따리 달성 퍼센트
         private int questionStateVal;                           //오늘의 퀴즈 상태
         private String questionStateMsg;                        //오늘의 퀴즈 상태메세지
         private QuestionDto.OrganizeQuestion todayQuiz;         //오늘의 퀴즈
